Added QUAPCInjectFakecmd
This allows us to create a process with a fake command line to better blend in. Again, this is just an example of how to extend TikiTorch that may be useful. I am open to other ideas too. In this case, we create a fake command line and this is sufficient, to evade detection, we don't really need to overwrite it with something, although you could.  See [Swamp Thing](https://github.com/FuzzySecurity/Sharp-Suite/tree/master/SwampThing) As an example. This is me being lazy. :)